refactor(terminal): extract effective-target helper to shared module
Hoists computeEffectiveKeyboardTarget and sameKeyboardFocusTarget into
src/shared/terminal-keyboard-target.ts as the single source of truth, replacing
the inline duplicate in terminal-debug-diagnostics.ts (back under the 500-line
hard cap) and the main-side definition. main + shared diagnostics now import
the shared helper. Includes lint:fix import/key ordering.
