+++ src/shared/contracts/terminal.ts
@@ -38,10 +38,10 @@ export type TerminalKeyboardFocusTarget =
 export interface TerminalInputRoutingSnapshot {
   /** dockview 活跃面板意图（terminal | web）。 */
   basePanel: TerminalKeyboardFocusTarget;
-  /** 当前活跃的浮层 web 焦点请求数；>0 即 effective=web。 */
-  webRequestCount: number;
   rendererSequence: number;
   webOverlayRects: TerminalWebOverlayRect[];
+  /** 当前活跃的浮层 web 焦点请求数；>0 即 effective=web。 */
+  webRequestCount: number;
 }
 
 export interface TerminalNativeInputRoutingSnapshot
